+++ lib/ktoast.dart
@@ -1,3 +1,3 @@
 library ktoast;
 
-export 'src/toast.dart' show showToast, KToast;
+export 'src/toast.dart' show showToast, KToast, ToastPosition;