+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## [0.0.3] add the toast position
+
+the toast can set position
+
+1.  showToast
+2.  defaultPosition
+
 ## [0.0.2] - add toast style
 
 add the textStyle